performance tweak on install song
