+++ app/index.ts
@@ -109,8 +109,7 @@ const electronApp = new IndexElectron(app, [
     join(loaderRootDir, 'cache.loader'),
     join(loaderRootDir, 'settings.loader'),
     join(loaderRootDir, 'played-maps.loader'),
-    join(loaderRootDir, 'installed-maps.loader'),
-    join(loaderRootDir, 'install.loader')
+    join(loaderRootDir, 'local-maps.loader')
 ]);
 
 electronApp
